Add ctrl-\ signal handling
