+++ .gitignore
@@ -8,3 +8,5 @@
 .gdb_history
 *.swp
 minishell
+test.sh
+minishell_test.log
